pyproject.toml
[tool.poetry]
name = "mini_embed"
version = "0.1.0"
description = ""
authors = ["exotol <artem.bardakov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.2"
pandas = "^1.4.2"
numpy = "^1.22.3"
torch = {path = "torch-1.11.0+cu113-cp39-cp39-win_amd64.whl"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyterlab = "^3.4.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
black = "^22.3.0"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
